Optimise find world location in overseas passports
This code seemed to find a world location and then checked to see if it
should find an alternate world location and called find world location
again when true.

The implementation `ALT_EMBASSIES.has_key?(value)` is replaced in favour
of `ALT_EMBASSIES[value] || value`, removing the very small possibility
of having an alternate embassy location key with a `nil` or `false`
value.