Remove `data_query` instance from outcome template
Extend the overseas passports calculator to accept a data_query
instance via the constructor, or create one if not received.

Add method `cash_only_country?` to calculator, which delegates to
`data_query.cash_only_countries?` to provide slightly clearer
method name, and to allow us to replace references to `data_query`
in the outcomes and partials.