Remove `location` variable from flow
Replace `location` variable in flow now available as `world_location` on
the calculator class, and add a validate block to ensure that the user
can't continue if they enter an invalid response.

Returning `nil` from validate keeps the user on the same question, which
is the same behaviour as raising `InvalidResponse` which is rescued and
keeps the user on the same question.