Define Marriage outcomes for Dominican Republic
This commit centres on both opposite and same sex marriages in the
Dominican Republic. This outcome seeks to isolate the outcomes for opposite
and same sex from other due to the increased paths and possible future changes
that could occur.

Here it starts with an if statement for the page title, followed by the body tag.
Within the body tag decisions are the made based on the rules given to the
developer.